Add support for sorting direction to the MediaManager (#3007)
* Remove redundant returns

`$this->putSession()` in `Backend\Traits\SessionMaker` doesn't return anything, and the return values of these methods are not used; so returning the value of `putSession()` in these methods is redundant.

* Add support for sort direction to MediaLibrary

Adds support for switching the direction that item lists get sorted by the media library. This necessitates the changing of the default direction of Last Modified and Size to Ascending, previously their sorting direction was Descending while Title was sorted in Ascending order. This shouldn't change anything as the user still has the power to sort by either direction. Fully backwards compatible with any calls to `sortItemList`.

* Support sorting direction in the MediaManager widget

Adds further support for specifying a sorting direction to the MediaManager widget itself

* Display sorting direction options

* Improving styles for sorting direction

* Send sorting information to the MediaManager

* Add translations for sorting feature

* Removing strange ghost whitespace